refactor: rename oclif-commands back to commands, fix tests
- Rename src/oclif-commands/ to src/commands/
- Use temp state file in task list tests to reduce output noise
- Remove alias test for config command
- Use this.log() instead of console.log() in task list command
