+++ src/commands/doctor.ts
