+++ src/commands/install.ts
