+++ src/commands/journal/utils.ts
