+++ src/commands/observe/report.ts
