+++ src/commands/observe/setup.ts
