+++ src/commands/observe/status.ts
