+++ src/commands/ralph/lib/formatter.ts
