+++ src/commands/ralph/marker/create.ts
@@ -8,8 +8,6 @@ import { generateMarker, MARKER_PREFIX } from '../lib/marker.js'
 export default class MarkerCreate extends BaseCommand {
   static override description = 'Generate a random marker for session tracking'
 
-  static override aliases = ['ralph:marker:create']
-
   static override examples = [
     '<%= config.bin %> ralph marker create',
   ]
